requirements.txt
fastapi
uvicorn[standard]
python-dotenv
redis
pydantic
pydantic-settings
email-validator
supabase
httpx
